translator/pyproject.toml
[project]
name = "clisp-translator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime==4.13.0",
    "black>=25.9.0",
    "jinja2>=3.1.6",
]

[tool.black]
line-length = 88
